Refactor cache analysis phase 6 helpers
