+++ src/pflow/core/prompt_cache_analysis/stages/row_builder.py
@@ -7,6 +7,7 @@ from collections.abc import Mapping
 from dataclasses import dataclass, replace
 from typing import Any
 
+from pflow.core.diagnostic import Diagnostic
 from pflow.core.llm_capabilities import get_min_cache_tokens
 from pflow.core.llm_config import get_default_workflow_model
 from pflow.core.llm_usage import normalize_litellm_usage_tokens
@@ -29,6 +30,7 @@ from ..token_estimation import (
 from ..token_estimation import (
     extract_unique_refs as _extract_unique_refs,
 )
+from ..trace_loading import _build_call_counts_by_node
 from ..types import (
     _BLOCK_ABSENT_BRANCH,
     _BLOCK_BELOW_PROVIDER_MIN,
@@ -38,9 +40,11 @@ from ..types import (
     CacheProjectionComponent,
     CrossWorkflowInputContribution,
     PerCallRow,
+    TraceExecutionIndex,
     _projection_component,
     _projection_source_confidence,
     _provider_min_state,
+    _RowCrossWorkflowCandidate,
     _safe_pct,
     aggregate_projection,
 )
@@ -59,6 +63,15 @@ class _PromptStaticTailFinding:
     meets_provider_min: bool | None = None
     provider_min_tokens: int | None = None
     blocked_reason: str = ""
+
+
+@dataclass(frozen=True)
+class _PerCallRowsResult:
+    rows: list[PerCallRow]
+    warnings: list[Diagnostic]
+    call_counts_by_node: dict[tuple[str | None, str], int]
+    cross_workflow_candidates_by_row: dict[tuple[str | None, str], list[_RowCrossWorkflowCandidate]]
+    has_greenfield_cross_workflow_projection_gap: bool = False
 
 
 def _total_observed_invocations(
@@ -800,6 +813,94 @@ def _build_cache_projection_components(
     )
 
 
+def _build_per_call_rows_and_warnings(
+    *,
+    ctx: AnalysisContext,
+    cw_result: Any,
+    trace_index: TraceExecutionIndex,
+) -> _PerCallRowsResult:
+    """Walk every reachable workflow IR and build LLM rows."""
+    from .cross_workflow import (
+        _build_cross_workflow_candidates_by_row,
+        _has_structural_cross_workflow_projection_candidate,
+    )
+    from .warnings import _per_node_warnings
+
+    rows: list[PerCallRow] = []
+    warnings: list[Diagnostic] = []
+    call_counts_by_node = _build_call_counts_by_node(ctx, cw_result)
+    cross_workflow_candidates_by_row = _build_cross_workflow_candidates_by_row(
+        ctx=ctx,
+        cw_result=cw_result,
+        call_counts_by_node=call_counts_by_node,
+        trace_index=trace_index,
+    )
+    has_greenfield_projection_gap = (
+        ctx.trace is None
+        and bool(getattr(cw_result, "edges", ()) or ())
+        and not cross_workflow_candidates_by_row
+        and _has_structural_cross_workflow_projection_candidate(cw_result)
+    )
+    for workflow_path, workflow_ir in getattr(cw_result, "irs_by_workflow", {}).items():
+        declared_chunks = _extract_declared_chunks(workflow_ir.get("cache"))
+        candidate_subsets_by_node = _detect_candidate_subsets(workflow_ir)
+        wf_ctx = AnalysisContext.build(
+            workflow_ir=workflow_ir,
+            parameters=ctx.parameters_for_workflow(workflow_path),
+            memo_cache=ctx.memo_cache,
+            trace_data=ctx.trace_data,
+            trace_outputs_by_key=ctx.trace_outputs_by_key,
+            workflow_path=workflow_path,
+            base_path=ctx.base_path,
+            parameters_by_workflow=ctx.parameters_by_workflow,
+            predicted_cache_keys=ctx.predicted_cache_keys,
+            prediction_fidelity_notes=ctx.prediction_fidelity_notes,
+            stale_memo_skipped=ctx.stale_memo_skipped,
+            stale_memo_uncheckable=ctx.stale_memo_uncheckable,
+        )
+        nodes = workflow_ir.get("nodes", []) or []
+        nodes_by_id: dict[str, dict[str, Any]] = {
+            str(n.get("id", "")): n for n in nodes if isinstance(n, dict) and n.get("id")
+        }
+        for node in nodes:
+            if not isinstance(node, dict) or node.get("type") != "llm":
+                continue
+            node_id = str(node.get("id", "?"))
+            row = _build_per_call_row(
+                node=node,
+                ctx=wf_ctx,
+                declared_chunks=declared_chunks,
+                candidate_subset=candidate_subsets_by_node.get(node_id),
+                trace_cost=trace_index.costs_by_key.get((workflow_path, node_id)),
+                trace_llm_call=trace_index.llm_calls_by_key.get((workflow_path, node_id)),
+                trace_llm_calls=trace_index.llm_call_lists_by_key.get((workflow_path, node_id), ()),
+                provider_trace_llm_call=trace_index.provider_llm_calls_by_key.get((workflow_path, node_id)),
+                provider_trace_llm_calls=trace_index.provider_llm_call_lists_by_key.get((workflow_path, node_id), ()),
+                did_not_execute_in_trace=(
+                    trace_index.trace_loaded and (workflow_path, node_id) not in trace_index.executed_keys
+                ),
+                cross_workflow_candidates_by_row=cross_workflow_candidates_by_row,
+            )
+            rows.append(row)
+            if not row.did_not_execute_in_trace:
+                warnings.extend(
+                    _per_node_warnings(
+                        node,
+                        row,
+                        declared_chunks=declared_chunks,
+                        nodes_by_id=nodes_by_id,
+                        ctx=wf_ctx,
+                    )
+                )
+    return _PerCallRowsResult(
+        rows=rows,
+        warnings=warnings,
+        call_counts_by_node=call_counts_by_node,
+        cross_workflow_candidates_by_row=cross_workflow_candidates_by_row,
+        has_greenfield_cross_workflow_projection_gap=has_greenfield_projection_gap,
+    )
+
+
 def _resolve_effective_row_model(explicit: Any, observed_models: tuple[str, ...]) -> tuple[str, bool]:
     """Return ``(model, model_is_heterogeneous)`` for one per-call row."""
     # Trace truth wins when present and unambiguous, because pricing,
@@ -986,6 +1087,61 @@ def _resolve_prompt_for_tokenization(prompt: str, ctx: AnalysisContext, node: di
 
     has_unresolved = bool(template_resolver_cls.TEMPLATE_PATTERN.search(resolved))
     return resolved, has_unresolved
+
+
+def _extract_declared_chunks(cache_block: Any) -> list[str]:
+    """Extract chunk names from a workflow's ``## Cache`` block IR."""
+    if not isinstance(cache_block, dict):
+        return []
+    items = cache_block.get("items") or []
+    if not isinstance(items, list):
+        return []
+    return [item.get("name", "") for item in items if isinstance(item, dict) and item.get("name")]
+
+
+def _detect_candidate_subsets(workflow_ir: dict[str, Any]) -> dict[str, list[str]]:
+    """Map LLM node_id to shared template refs used by at least two nodes."""
+    if _extract_declared_chunks(workflow_ir.get("cache")):
+        return {}
+    ref_to_nodes = _collect_candidate_template_refs(workflow_ir)
+    candidates_by_node: dict[str, list[str]] = {}
+    for ref, node_ids in ref_to_nodes.items():
+        if len(node_ids) < 2:
+            continue
+        for node_id in node_ids:
+            candidates_by_node.setdefault(node_id, []).append(ref)
+    return candidates_by_node
+
+
+def _collect_candidate_template_refs(workflow_ir: dict[str, Any]) -> dict[str, list[str]]:
+    """Collect non-batch LLM prompt refs by the nodes that consume them."""
+    ref_to_nodes: dict[str, list[str]] = {}
+    for node in workflow_ir.get("nodes", []) or []:
+        if not isinstance(node, dict) or node.get("type") != "llm":
+            continue
+        prompt = node.get("params", {}).get("prompt", "")
+        if not isinstance(prompt, str):
+            continue
+        batch_aliases = _batch_aliases(node)
+        seen_in_node: set[str] = set()
+        for classified_ref in classify_prompt_refs(prompt, batch_alias=None, node_inputs=_node_inputs(node)):
+            for ref in classified_ref.operand_paths:
+                if _is_batch_scoped_ref(ref, batch_aliases) or ref in seen_in_node:
+                    continue
+                seen_in_node.add(ref)
+                ref_to_nodes.setdefault(ref, []).append(str(node["id"]))
+    return ref_to_nodes
+
+
+def _batch_aliases(node: dict[str, Any]) -> set[str]:
+    batch = node.get("batch")
+    if not isinstance(batch, dict):
+        return set()
+    return {str(batch.get("as", "item"))}
+
+
+def _is_batch_scoped_ref(ref: str, aliases: set[str]) -> bool:
+    return any(ref == alias or ref.startswith(f"{alias}.") or ref.startswith(f"{alias}[") for alias in aliases)
 
 
 def _node_inputs(node: dict[str, Any]) -> Mapping[str, Any] | None:
